refactor: Add DefaultLayout component
